+++ internal/domain/claims.go
@@ -11,11 +11,11 @@ type CustomClaims struct {
 	jwt.RegisteredClaims
 }
 
-func (c *CustomClaims) UserID() uint {
+func (c *CustomClaims) UserID() int64 {
 	sub := c.Subject
 	res, err := strconv.Atoi(sub)
 	if err != nil {
 		return 0
 	}
-	return uint(res)
+	return int64(res)
 }
