+++ internal/domain/question.go
@@ -1,8 +1,8 @@
 package domain
 
 type Question struct {
-	ID          uint
-	TopicID     uint
+	ID          int64
+	TopicID     int64
 	Text        string
 	Code        string
 	Explanation string
