+++ internal/domain/taskRepo.go
@@ -2,6 +2,6 @@ package domain
 
 type TaskRepository interface {
 	GetAll() ([]Task, error)
-	GetAllByUserID(ID uint) ([]Task, error)
-	UpdateCompleted(taskID, userID uint, completed bool) error
+	GetAllByUserID(ID int64) ([]Task, error)
+	UpdateCompleted(taskID, userID int64, completed bool) error
 }
