+++ internal/domain/topicRepo.go
@@ -3,5 +3,5 @@ package domain
 type TopicRepository interface {
 	All() ([]*Topic, error)
 	AllWithCount() ([]TopicWithCount, error)
-	GetByID(id uint) (*Topic, error)
+	GetByID(id int64) (*Topic, error)
 }
