+++ internal/domain/userRepo.go
@@ -2,9 +2,9 @@ package domain
 
 type UserRepository interface {
 	CreateUser(user User) error
-	DeleteUser(id uint) error
+	DeleteUser(id int64) error
 	GetByLogin(login string) (User, error)
-	GetByID(id uint) (User, error)
-	GetByIDs(ids []uint) ([]User, error)
+	GetByID(id int64) (User, error)
+	GetByIDs(ids []int64) ([]User, error)
 	GetAll() ([]User, error)
 }
