+++ internal/infra/persistence/gorm/messageEntity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type MessageEntity struct {
-	ID        uint      `gorm:"primaryKey"`
-	UserID    uint      `gorm:"column:user_id"`
+	ID        int64     `gorm:"primaryKey"`
+	UserID    int64     `gorm:"column:user_id"`
 	Content   string    `gorm:"column:content"`
 	Type      string    `gorm:"column:message_type"`
 	CreatedAt time.Time `gorm:"column:created_at"`
