+++ internal/infra/persistence/gorm/questionEntity.go
@@ -1,8 +1,8 @@
 package gorm
 
 type QuestionEntity struct {
-	ID          uint           `gorm:"primaryKey"`
-	TopicID     uint           `gorm:"column:topic_id"`
+	ID          int64          `gorm:"primaryKey"`
+	TopicID     int64          `gorm:"column:topic_id"`
 	Text        string         `gorm:"column:text"`
 	Code        string         `gorm:"column:code"`
 	Explanation string         `gorm:"column:explanation"`
