+++ internal/infra/persistence/gorm/taskEntity.go
@@ -3,14 +3,14 @@ package gorm
 import "time"
 
 type TaskEntity struct {
-	ID          uint   `gorm:"primaryKey"`
+	ID          int64  `gorm:"primaryKey"`
 	Name        string `gorm:"not null"`
 	Description string
 	CreatedAt   time.Time `gorm:"autoCreateTime"`
 	UpdatedAt   time.Time `gorm:"autoUpdateTime"`
 	Completed   bool      `gorm:"default:false"`
 
-	UserID uint
+	UserID int64
 }
 
 func (TaskEntity) TableName() string {
